avec EtudiantFacrtory

+++ 2230912_2130331_Lab5Partie2/2230912_2130331_Lab5Partie2/Models/Cours.cs
@@ -27,7 +27,7 @@
 
         }
 
-        public Cours (string codeCours, string titreCours, double dureeCours)
+        public Cours (string codeCours, string titreCours, int dureeCours)
         {
             CodeCours = codeCours;
             TitreCours = titreCours;
